fix: update dpki api name
