+++ cmd/client/commands/dpki_oracle.go
@@ -2,7 +2,6 @@ package commands
 
 import (
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"github.com/abiosoft/ishell"
 	"github.com/qlcchain/go-qlc/cmd/util"
@@ -12,7 +11,7 @@ import (
 	rpc "github.com/qlcchain/jsonrpc2"
 )
 
-func addPublishCmdByShell(parentCmd *ishell.Cmd) {
+func addOraclePublishCmdByShell(parentCmd *ishell.Cmd) {
 	account := util.Flag{
 		Name:  "account",
 		Must:  true,
@@ -37,23 +36,23 @@ func addPublishCmdByShell(parentCmd *ishell.Cmd) {
 		Usage: "publish public key",
 		Value: "",
 	}
-	fee := util.Flag{
-		Name:  "fee",
+	code := util.Flag{
+		Name:  "code",
 		Must:  true,
-		Usage: "publish fee (at least 5 QGAS)",
+		Usage: "verification code",
 		Value: "",
 	}
-	verifiers := util.Flag{
-		Name:  "verifiers",
+	hash := util.Flag{
+		Name:  "hash",
 		Must:  true,
-		Usage: "verifiers",
+		Usage: "hash verified for",
 		Value: "",
 	}
 	c := &ishell.Cmd{
-		Name: "publish",
-		Help: "publish id and key",
+		Name: "oracle",
+		Help: "oracle publish id and key",
 		Func: func(c *ishell.Context) {
-			args := []util.Flag{account, typ, id, pk, fee, verifiers}
+			args := []util.Flag{account, typ, id, pk, code, hash}
 			if util.HelpText(c, args) {
 				return
 			}
@@ -67,10 +66,10 @@ func addPublishCmdByShell(parentCmd *ishell.Cmd) {
 			typeP := util.StringVar(c.Args, typ)
 			idP := util.StringVar(c.Args, id)
 			pkP := util.StringVar(c.Args, pk)
-			feeP := util.StringVar(c.Args, fee)
-			verifiersP := util.StringVar(c.Args, verifiers)
+			codeP := util.StringVar(c.Args, code)
+			hashP := util.StringVar(c.Args, hash)
 
-			err := publish(accountP, typeP, idP, pkP, feeP, verifiersP)
+			err := oraclePublish(accountP, typeP, idP, pkP, codeP, hashP)
 			if err != nil {
 				util.Warn(err)
 			}
@@ -79,7 +78,7 @@ func addPublishCmdByShell(parentCmd *ishell.Cmd) {
 	parentCmd.AddCmd(c)
 }
 
-func publish(accountP, typeP, idP, pkP, feeP, verifiersP string) error {
+func oraclePublish(accountP, typeP, idP, pkP, codeP, hashP string) error {
 	if accountP == "" {
 		return fmt.Errorf("account can not be null")
 	}
@@ -96,12 +95,12 @@ func publish(accountP, typeP, idP, pkP, feeP, verifiersP string) error {
 		return fmt.Errorf("publish public key can not be null")
 	}
 
-	if feeP == "" {
-		return fmt.Errorf("publish fee can not be null")
+	if codeP == "" {
+		return fmt.Errorf("verification can not be null")
 	}
 
-	if verifiersP == "" {
-		return fmt.Errorf("verifiers can not be null")
+	if hashP == "" {
+		return fmt.Errorf("verified hash can not be null")
 	}
 
 	accBytes, err := hex.DecodeString(accountP)
@@ -120,37 +119,29 @@ func publish(accountP, typeP, idP, pkP, feeP, verifiersP string) error {
 	}
 	defer client.Close()
 
-	verifiers := make([]types.Address, 0)
-	err = json.Unmarshal([]byte(verifiersP), &verifiers)
-	if err != nil {
-		return err
+	param := &api.OracleParam{
+		Account: acc.Address(),
+		OType:   typeP,
+		OID:     idP,
+		PubKey:  pkP,
+		Code:    codeP,
+		Hash:    hashP,
 	}
 
-	param := &api.PublishParam{
-		Account:   acc.Address(),
-		PType:     typeP,
-		PID:       idP,
-		PubKey:    pkP,
-		Fee:       types.StringToBalance(feeP),
-		Verifiers: verifiers,
-	}
-
-	var blockInfo api.PublishRet
-	err = client.Call(&blockInfo, "pkd_getPublishBlock", param)
+	var block types.StateBlock
+	err = client.Call(&block, "dpki_getOracleBlock", param)
 	if err != nil {
 		return err
 	}
 
 	var w types.Work
-	block := blockInfo.Block
 	worker, _ := types.NewWorker(w, block.Root())
 	block.Work = worker.NewWork()
 
 	hash := block.GetHash()
 	block.Signature = acc.Sign(hash)
 
-	fmt.Printf("publish block:\n%s\nhash[%s]\nverifiers:\n%s\n", cutil.ToIndentString(block), block.GetHash(),
-		cutil.ToIndentString(blockInfo.Verifiers))
+	fmt.Printf("oracle block:\n%s\nhash[%s]\n", cutil.ToIndentString(block), block.GetHash())
 
 	var h types.Hash
 	err = client.Call(&h, "ledger_process", &block)
