+++ cmd/client/commands/dpki_publish.go
@@ -2,6 +2,7 @@ package commands
 
 import (
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
 	"github.com/abiosoft/ishell"
 	"github.com/qlcchain/go-qlc/cmd/util"
@@ -11,7 +12,7 @@ import (
 	rpc "github.com/qlcchain/jsonrpc2"
 )
 
-func addUnPublishCmdByShell(parentCmd *ishell.Cmd) {
+func addPublishCmdByShell(parentCmd *ishell.Cmd) {
 	account := util.Flag{
 		Name:  "account",
 		Must:  true,
@@ -21,20 +22,38 @@ func addUnPublishCmdByShell(parentCmd *ishell.Cmd) {
 	typ := util.Flag{
 		Name:  "type",
 		Must:  true,
-		Usage: "unPublish id type (email/weChat)",
+		Usage: "publish id type (email/weChat)",
 		Value: "",
 	}
 	id := util.Flag{
 		Name:  "id",
 		Must:  true,
-		Usage: "unPublish id (email address/weChat id)",
+		Usage: "publish id (email address/weChat id)",
+		Value: "",
+	}
+	pk := util.Flag{
+		Name:  "pk",
+		Must:  true,
+		Usage: "publish public key",
+		Value: "",
+	}
+	fee := util.Flag{
+		Name:  "fee",
+		Must:  true,
+		Usage: "publish fee (at least 5 QGAS)",
+		Value: "",
+	}
+	verifiers := util.Flag{
+		Name:  "verifiers",
+		Must:  true,
+		Usage: "verifiers",
 		Value: "",
 	}
 	c := &ishell.Cmd{
-		Name: "unPublish",
-		Help: "unPublish id and key",
+		Name: "publish",
+		Help: "publish id and key",
 		Func: func(c *ishell.Context) {
-			args := []util.Flag{account, typ, id}
+			args := []util.Flag{account, typ, id, pk, fee, verifiers}
 			if util.HelpText(c, args) {
 				return
 			}
@@ -47,8 +66,11 @@ func addUnPublishCmdByShell(parentCmd *ishell.Cmd) {
 			accountP := util.StringVar(c.Args, account)
 			typeP := util.StringVar(c.Args, typ)
 			idP := util.StringVar(c.Args, id)
+			pkP := util.StringVar(c.Args, pk)
+			feeP := util.StringVar(c.Args, fee)
+			verifiersP := util.StringVar(c.Args, verifiers)
 
-			err := unPublish(accountP, typeP, idP)
+			err := publish(accountP, typeP, idP, pkP, feeP, verifiersP)
 			if err != nil {
 				util.Warn(err)
 			}
@@ -57,7 +79,7 @@ func addUnPublishCmdByShell(parentCmd *ishell.Cmd) {
 	parentCmd.AddCmd(c)
 }
 
-func unPublish(accountP, typeP, idP string) error {
+func publish(accountP, typeP, idP, pkP, feeP, verifiersP string) error {
 	if accountP == "" {
 		return fmt.Errorf("account can not be null")
 	}
@@ -68,6 +90,18 @@ func unPublish(accountP, typeP, idP string) error {
 
 	if idP == "" {
 		return fmt.Errorf("publish id can not be null")
+	}
+
+	if pkP == "" {
+		return fmt.Errorf("publish public key can not be null")
+	}
+
+	if feeP == "" {
+		return fmt.Errorf("publish fee can not be null")
+	}
+
+	if verifiersP == "" {
+		return fmt.Errorf("verifiers can not be null")
 	}
 
 	accBytes, err := hex.DecodeString(accountP)
@@ -86,26 +120,37 @@ func unPublish(accountP, typeP, idP string) error {
 	}
 	defer client.Close()
 
-	param := &api.UnPublishParam{
-		Account: acc.Address(),
-		PType:   typeP,
-		PID:     idP,
+	verifiers := make([]types.Address, 0)
+	err = json.Unmarshal([]byte(verifiersP), &verifiers)
+	if err != nil {
+		return err
 	}
 
-	var block types.StateBlock
-	err = client.Call(&block, "pkd_getUnPublishBlock", param)
+	param := &api.PublishParam{
+		Account:   acc.Address(),
+		PType:     typeP,
+		PID:       idP,
+		PubKey:    pkP,
+		Fee:       types.StringToBalance(feeP),
+		Verifiers: verifiers,
+	}
+
+	var blockInfo api.PublishRet
+	err = client.Call(&blockInfo, "dpki_getPublishBlock", param)
 	if err != nil {
 		return err
 	}
 
 	var w types.Work
+	block := blockInfo.Block
 	worker, _ := types.NewWorker(w, block.Root())
 	block.Work = worker.NewWork()
 
 	hash := block.GetHash()
 	block.Signature = acc.Sign(hash)
 
-	fmt.Printf("unPublish block:\n%s\nhash[%s]\n", cutil.ToIndentString(block), block.GetHash())
+	fmt.Printf("publish block:\n%s\nhash[%s]\nverifiers:\n%s\n", cutil.ToIndentString(block), block.GetHash(),
+		cutil.ToIndentString(blockInfo.Verifiers))
 
 	var h types.Hash
 	err = client.Call(&h, "ledger_process", &block)
