+++ cmd/client/commands/dpki_verifier_register.go
@@ -11,7 +11,7 @@ import (
 	rpc "github.com/qlcchain/jsonrpc2"
 )
 
-func addVerifierUnRegisterCmdByShell(parentCmd *ishell.Cmd) {
+func addVerifierRegisterCmdByShell(parentCmd *ishell.Cmd) {
 	account := util.Flag{
 		Name:  "account",
 		Must:  true,
@@ -24,11 +24,17 @@ func addVerifierUnRegisterCmdByShell(parentCmd *ishell.Cmd) {
 		Usage: "verifier type(email/weChat)",
 		Value: "",
 	}
+	vInfo := util.Flag{
+		Name:  "info",
+		Must:  true,
+		Usage: "verifiers address(email address/weChat ID)",
+		Value: "",
+	}
 	c := &ishell.Cmd{
-		Name: "unregister",
-		Help: "unregister verifier",
+		Name: "register",
+		Help: "register verifier",
 		Func: func(c *ishell.Context) {
-			args := []util.Flag{account, vType}
+			args := []util.Flag{account, vType, vInfo}
 			if util.HelpText(c, args) {
 				return
 			}
@@ -40,8 +46,9 @@ func addVerifierUnRegisterCmdByShell(parentCmd *ishell.Cmd) {
 
 			accountP := util.StringVar(c.Args, account)
 			vTypeP := util.StringVar(c.Args, vType)
+			vInfoP := util.StringVar(c.Args, vInfo)
 
-			err := verifierUnRegister(accountP, vTypeP)
+			err := verifierRegister(accountP, vTypeP, vInfoP)
 			if err != nil {
 				util.Warn(err)
 			}
@@ -50,13 +57,17 @@ func addVerifierUnRegisterCmdByShell(parentCmd *ishell.Cmd) {
 	parentCmd.AddCmd(c)
 }
 
-func verifierUnRegister(accountP, vTypeP string) error {
+func verifierRegister(accountP, vTypeP, vInfoP string) error {
 	if accountP == "" {
 		return fmt.Errorf("account can not be null")
 	}
 
 	if vTypeP == "" {
 		return fmt.Errorf("verifier type can not be null")
+	}
+
+	if vInfoP == "" {
+		return fmt.Errorf("verifier info can not be null")
 	}
 
 	accBytes, err := hex.DecodeString(accountP)
@@ -75,13 +86,14 @@ func verifierUnRegister(accountP, vTypeP string) error {
 	}
 	defer client.Close()
 
-	param := &api.VerifierUnRegParam{
+	param := &api.VerifierRegParam{
 		Account: acc.Address(),
 		VType:   vTypeP,
+		VInfo:   vInfoP,
 	}
 
 	var block types.StateBlock
-	err = client.Call(&block, "pkd_getVerifierUnregisterBlock", param)
+	err = client.Call(&block, "dpki_getVerifierRegisterBlock", param)
 	if err != nil {
 		return err
 	}
@@ -93,7 +105,7 @@ func verifierUnRegister(accountP, vTypeP string) error {
 	hash := block.GetHash()
 	block.Signature = acc.Sign(hash)
 
-	fmt.Printf("unregister block:\n%s\nhash[%s]\n", cutil.ToIndentString(block), block.GetHash())
+	fmt.Printf("register block:\n%s\nhash[%s]\n", cutil.ToIndentString(block), block.GetHash())
 
 	var h types.Hash
 	err = client.Call(&h, "ledger_process", &block)
